+++ NEWS.md
@@ -1,8 +1,11 @@
 # 📰 Nudebomb News
 
-## v0.5.0
+## v1.0.0
 
+- Fetch media file language from tmdb. You don't have to use lang files if your
+  media is in tmdb.
 - Relabel und tracks with a preferred language option.
+- Don't write timestamps if nothing stripped.
 
 ## v0.4.9